fixed some bugs with the display
