+++ Assets/Prefabs/ElvisUnit.prefab
@@ -43,6 +43,7 @@ GameObject:
   m_Component:
   - 4: {fileID: 474714}
   - 114: {fileID: 11436486}
+  - 114: {fileID: 11473582}
   m_Layer: 8
   m_Name: ElvisUnit
   m_TagString: Untagged
@@ -237,17 +238,17 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 813fb0e327e8047a985198d4de57843d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  HitPoints: 0
-  AttackRange: 0
-  AttackFactor: 0
-  DefenceFactor: 0
-  MovementPoints: 0
-  MovementSpeed: 0
-  ActionPoints: 0
+  HitPoints: 100
+  AttackRange: 3
+  AttackFactor: 100
+  DefenceFactor: 50
+  MovementPoints: 4
+  MovementSpeed: 8
+  ActionPoints: 1
   PlayerNumber: 1
   ColorButton: 3
   unitColor: {r: 0.12282246, g: 0.007298882, b: 0.99264705, a: 0.991}
-  hopHeight: 10
+  hopHeight: 1
   character: 0
 --- !u!114 &11448352
 MonoBehaviour:
@@ -270,6 +271,19 @@ MonoBehaviour:
   m_FallbackScreenDPI: 96
   m_DefaultSpriteDPI: 96
   m_DynamicPixelsPerUnit: 1
+--- !u!114 &11473582
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 163780}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 0d24c4413c92fb64297372a1991c9af3, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  DisableTurns: 5
+  JailPrefab: {fileID: 152296, guid: 965e326b5f1c17f428dd5dbe5d1b0c92, type: 2}
 --- !u!124 &12441814
 Behaviour:
   m_ObjectHideFlags: 1
