+++ Assets/scripts/ElvisUnit.cs
@@ -2,17 +2,5 @@
 using System.Collections;
 
 public class ElvisUnit : MelodyUnit {
-
-	// Use this for initialization
-	void Start () {
-		HitPoints = 100;
-		AttackRange = 3;
-		AttackFactor = 100;
-		DefenceFactor = 50;
-		MovementPoints = 4;
-		MovementSpeed = 8;
-		ActionPoints = 1;
-		hopHeight = 1;
-	}
 	
 }
