+++ Assets/scripts/MileyUnit.cs
@@ -3,16 +3,4 @@ using System.Collections;
 
 public class MileyUnit : MelodyUnit {
 
-	// Use this for initialization
-	void Start () {
-		HitPoints = 100;
-		AttackRange = 3;
-		AttackFactor = 100;
-		DefenceFactor = 50;
-		MovementPoints = 4;
-		MovementSpeed = 8;
-		ActionPoints = 1;	
-		hopHeight = 1;	
-	}
-
 }
