+++ Assets/scripts/Spotlight.cs.meta
